Add batch message container builder.
