+++ pulsar-client-api/src/main/java/org/apache/pulsar/client/internal/DefaultImplementation.java
@@ -36,6 +36,7 @@ import java.util.function.Supplier;
 import lombok.experimental.UtilityClass;
 
 import org.apache.pulsar.client.api.Authentication;
+import org.apache.pulsar.client.api.BatchMessageContainerBuilder;
 import org.apache.pulsar.client.api.ClientBuilder;
 import org.apache.pulsar.client.api.MessageId;
 import org.apache.pulsar.client.api.Schema;
@@ -281,5 +282,11 @@ public class DefaultImplementation {
         return catchExceptions(
                 () -> (RecordSchemaBuilder) getConstructor("org.apache.pulsar.client.impl.schema.RecordSchemaBuilderImpl",
                         String.class).newInstance(name));
+    }
+
+    public static BatchMessageContainerBuilder newDefaultBatchMessageContainerBuilder() {
+        return catchExceptions(
+            () -> (BatchMessageContainerBuilder) getConstructor("org.apache.pulsar.client.impl.DefaultBatchMessageContainerBuilder")
+                    .newInstance());
     }
 }
