+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/ProducerBuilderImpl.java
@@ -29,6 +29,7 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.commons.lang3.StringUtils;
+import org.apache.pulsar.client.api.BatchMessageContainerBuilder;
 import org.apache.pulsar.client.api.CompressionType;
 import org.apache.pulsar.client.api.CryptoKeyReader;
 import org.apache.pulsar.client.api.HashingScheme;
@@ -222,6 +223,13 @@ public class ProducerBuilderImpl<T> implements ProducerBuilder<T> {
         conf.setBatchingMaxMessages(batchMessagesMaxMessagesPerBatch);
         return this;
     }
+
+    @Override
+    public ProducerBuilder<T> batchingContainerBuilder(BatchMessageContainerBuilder batchingContainerBuilder) {
+        conf.setBatchMessageContainerBuilder(batchingContainerBuilder);
+        return this;
+    }
+
 
     @Override
     public ProducerBuilder<T> initialSequenceId(long initialSequenceId) {
